Connect to light device when selected and support sending messages
Connects to light device when selected in the settings

Continues trying to reconnect if connection aborts or connection can't
be established

Parses node info message.

Made widget in upper right corner that shows connection status.
by howering over the widget it is possible to get device name and
network size. If no device selected, the connection status is hidden.

Implement support for sending messages. Currently it is possible to
send these message types:
* Light Mode
* Strobe
* Glitter
* Pulse
* Onboard LED strength

Set color to teamcolor and lightmode to on when a song is guessed
Turn off onboard LED when connected