Clean up quizloading in the QuizCreator.
Much of the loading functionality in QuizFactory has been split into
QuizData, EntryCreator and QuizCreator. The creators now has the responsibility
to construct itself from a boost property tree.

QuizData is now split into its own module and made a class that
does the loading of the categories and entries.

Added functionality to load an existing category in the quizCreator